JS code is divided into modules. Modules are linked via import/export. Main.js is connected as module to index.html. Functions.js is removed as unused.
