+++ js/const.js
@@ -1,0 +1,27 @@
+const NAMES = ['Александр', 'Леонид', 'Вячеслав', 'Сергей', 'Влад', 'Петр'];
+
+const MESSAGES = ['Все отлично!', 'В целом, всё неплохо. Но не всё.', 'Когда вы делаете фотографии, хорошо бы убирать палец из кадра. В конце концов, это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.', 'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+const DESCRIPTIONS = ['Отдыхаем от забот', 'Это фото - образец лаконичности', 'Заснять каждый момент', 'Любовь в каждом кадре', 'Фото, вошедшее в историю'];
+
+const NUMBER_OF_OBJECTS = 25;
+
+const NUMBER_OF_COMMENTS = 3;
+
+const MIN_ID = 1;
+
+const MAX_ID = 25;
+
+const MIN_URL = 1;
+
+const MAX_URL = 25;
+
+const MIN_LIKES = 15;
+
+const MAX_LIKES = 200;
+
+const MIN_AVATAR = 1;
+
+const MAX_AVATAR = 6;
+
+export {NAMES, MESSAGES, DESCRIPTIONS, NUMBER_OF_OBJECTS, NUMBER_OF_COMMENTS, MIN_ID, MAX_ID, MIN_URL, MAX_URL, MIN_LIKES, MAX_LIKES, MIN_AVATAR, MAX_AVATAR};
